requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
Flask-Limiter==3.3.0
Flask-CORS==4.0.0
python-dotenv==1.0.0